python-example/requirements.txt
flask==1.1.1
werkzeug==0.16.0
connexion[swagger-ui]==2.5.1
python_dateutil==2.6.0
python-dotenv==0.10.3
setuptools==40.8.0
swagger_ui_bundle==0.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4